[5기 - 이세희] SpringBoot Part2 Weekly Mission 제출합니다.

+++ src/main/java/com/programmers/vouchermanagement/voucher/dto/CreateVoucherRequest.java
@@ -1,8 +1,6 @@
 package com.programmers.vouchermanagement.voucher.dto;
 
-import java.math.BigDecimal;
-
 import com.programmers.vouchermanagement.voucher.domain.VoucherType;
 
-public record CreateVoucherRequest(BigDecimal discountValue, VoucherType voucherType) {
+public record CreateVoucherRequest(long discountValue, VoucherType voucherType) {
 }
